[refactor] #53 - 리마인드 설정 컴포넌트 분리
